Move decryptNotes perf job into a separate file
